pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flwr-app"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.16.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "wandb"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "Hawking"

[tool.flwr.app.components]
serverapp = "flwr_app.server_app:app"
clientapp = "flwr_app.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 5
fraction-fit = 0.5
local-epochs = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10
